CR adjustments
