--- apps/checkout-app/backend/payments/utils.ts
@@ -1,5 +1,0 @@
-export const parseAmountToString = (amount: number, negative = false) => {
-  const value = amount.toFixed(2).toString();
-
-  return negative ? "-" + value : value;
-};
